+++ pumpkin/src/block/blocks/fire/fire.rs
@@ -1,4 +1,4 @@
-﻿use pumpkin_data::biome::Biome;
+use pumpkin_data::biome::Biome;
 use pumpkin_data::block_properties::{BlockProperties, EnumVariants, HorizontalAxis};
 use pumpkin_data::dimension::Dimension;
 use pumpkin_data::fluid::Fluid;
@@ -176,7 +176,7 @@ impl FireBlock {
                 fire_props.age = EnumVariants::from_index(new_age);
                 let new_state_id = fire_props.to_state_id(&Block::FIRE);
                 if let Some(server) = world.server.upgrade() {
-                    let spread_event = crate::plugin::block::block_spread::BlockSpreadEvent {
+                    let spread_event = crate::plugin::block::spread::BlockSpreadEvent {
                         source_block: &Block::FIRE,
                         source_pos: *source_pos,
                         block: &Block::FIRE,
@@ -186,7 +186,7 @@ impl FireBlock {
                     };
                     let spread_event = server
                         .plugin_manager
-                        .fire::<crate::plugin::block::block_spread::BlockSpreadEvent>(spread_event)
+                        .fire::<crate::plugin::block::spread::BlockSpreadEvent>(spread_event)
                         .await;
                     if spread_event.cancelled {
                         return;
@@ -206,7 +206,7 @@ impl FireBlock {
                     .await;
             } else {
                 if let Some(server) = world.server.upgrade() {
-                    let spread_event = crate::plugin::block::block_spread::BlockSpreadEvent {
+                    let spread_event = crate::plugin::block::spread::BlockSpreadEvent {
                         source_block: &Block::FIRE,
                         source_pos: *source_pos,
                         block: &Block::FIRE,
@@ -216,7 +216,7 @@ impl FireBlock {
                     };
                     let spread_event = server
                         .plugin_manager
-                        .fire::<crate::plugin::block::block_spread::BlockSpreadEvent>(spread_event)
+                        .fire::<crate::plugin::block::spread::BlockSpreadEvent>(spread_event)
                         .await;
                     if spread_event.cancelled {
                         return;
@@ -558,7 +558,7 @@ impl BlockBehaviour for FireBlock {
 
                                     if let Some(server) = world.server.upgrade() {
                                         let spread_event =
-                                            crate::plugin::block::block_spread::BlockSpreadEvent {
+                                            crate::plugin::block::spread::BlockSpreadEvent {
                                                 source_block: &Block::FIRE,
                                                 source_pos: *pos,
                                                 block: &Block::FIRE,
@@ -568,7 +568,7 @@ impl BlockBehaviour for FireBlock {
                                             };
                                         let spread_event = server
                                             .plugin_manager
-                                            .fire::<crate::plugin::block::block_spread::BlockSpreadEvent>(
+                                            .fire::<crate::plugin::block::spread::BlockSpreadEvent>(
                                                 spread_event,
                                             )
                                             .await;
